requirements.txt
streamlit
scikit-learn
matplotlib.pyplot
mlflow
requests
pandas
apache-airflow==2.6.0